+++ lizard_portal/media/lizard_portal/lizard/form/BackgroundLayerSelector.js
@@ -17,6 +17,7 @@
           title: 'Achtergroundkaart selectie',
           is_background_selection: true,
           modal: true,
+          constrainHeader: true,
           items: Ext.create('Lizard.form.BackgroundLayerSelector', config)
         }).show();
       }