+++ lizard_portal/media/lizard_portal/lizard/popup/FeatureInfo.js
@@ -28,6 +28,7 @@
         return Ext.create('Ext.window.Window', {
           title: 'Info',
           popup_type: 'feature_info',
+          constrainHeader: true,
           items: [
             {
               xtype: 'panel',