+++ lizard_portal/media/lizard_portal/lizard/portlet/WorkspacePortlet.js
@@ -89,6 +89,7 @@
             modal: true,
             xtype: 'leditgrid',
             editpopup: true,
+            constrainHeader: true,
             items: [
               {
                 xtype: 'workspacesaveform',
@@ -112,6 +113,7 @@
             width: 600,
             height: window.innerHeight - 200,
             modal: true,
+            constrainHeader: true,
             layout: {
               type: 'vbox',
               align: 'stretch'