+++ lizard_portal/media/lizard_portal/lizard/window/Screen.js
@@ -69,7 +69,8 @@ showNavigationPortalTemplate
       window_settings = {
         title: title,
         width: 800,
-        height: 500,
+        height: 600,
+        constrainHeader: true,
         autoScroll: true,
         bodyStyle: {
           background: 'white'
@@ -93,6 +94,7 @@ showNavigationPortalTemplate
       if (window_options.save) {
         window_settings.tools.push({
           type: 'save',
+          tooltip: window_options.save[0],
           handler: function(e, target, panelHeader, tool) {
             return me.linkToPopup.apply(me, window_options.save);
           }
@@ -101,6 +103,7 @@ showNavigationPortalTemplate
       if (window_options.search) {
         window_settings.tools.push({
           type: 'search',
+          tooltip: window_options.search[0],
           handler: function(e, target, panelHeader, tool) {
             return me.linkToPopup.apply(me, window_options.search);
           }
@@ -109,6 +112,7 @@ showNavigationPortalTemplate
       if (window_options.plus) {
         window_settings.tools.push({
           type: 'plus',
+          tooltip: window_options.plus[0],
           handler: function(e, target, panelHeader, tool) {
             return me.linkToPopup.apply(me, window_options.plus);
           }