+++ lizard_portal/templates/portals/valideer_configuraties.js
@@ -75,14 +75,17 @@
 
 				Ext.create('Ext.window.Window', {
 					       title: 'Verschillen tussen configuraties',
+					       autoHeight: false,
+					       height: Ext.getBody().getViewSize().height * .50,
+					       autoScroll: true,
 					       width: 800,
-					       height: 600,
 					       modal: true,
 					       editpopup: false,
+				               constrainHeader: true,
 					       loader:{
 						   loadMask: true,
 						   autoLoad: true,
-						   url: '/portal/diff/' + record.data.polder,
+						   url: '/validation/diff/' + record.data.polder + '/' + record.data.type,
 						   ajaxOptions: {
 						       method: 'GET'
 						   },